added unit tests for entity POV export
the unit tests use a set of gold standard files that are then
compared to the output of gfx::Scene::Instance().ExportPov().
This assumes that there are no differences due to errors in
floating point arithmetic.

If floating point errors tend to show on some of the platforms,
we  will have find another way to check the correctness of the
POV export.

git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@1860 5a81b35b-ba03-0410-adc8-b2c5c5119f08