fix several problems in PDBWriter
Residues with names shorter than 3 letters are left-padded,
we throw an error if we encounter chain, residue and atom
names that are too long (longer than 1, 3 and 4 characters
respectively) and we try to put the atom name into the
correct column.

Oh, and we insert TER records at the correct positions.

All in all the code is much cleaner and faster as it does
not longer rely on boost.format (yes, io performance is
sometimes critical).

git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@2215 5a81b35b-ba03-0410-adc8-b2c5c5119f08