fix problem when deleting selection with backspace
When backspace is pressed and the cursor position is equal to the
start of the edit block, the selected text was not deleted as it
should. Instead of testing for textCursor().position() we now use
textCursor().selectionStart() which is always equal to the start
of the selection. In case no text is selected, it is equal to
position().

Andi, please review.

git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@2381 5a81b35b-ba03-0410-adc8-b2c5c5119f08