implement deployment mechanism Leopard
Leopard's install_name_tool doesn't allow to manipulate the
rpaths of mach-o binaries, that's why we have to fall back to the
ugly @executable_path on these systems. This doesn't work properly
when using the default Python coming with the system, since the
library load commands are defined relative to the executable they
were loaded in. If the python the binary and gosty binary are not
in the same path, BANG!

git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@2479 5a81b35b-ba03-0410-adc8-b2c5c5119f08