fix compilation on MacOS X with shaders enabled
MacOS X doesn't require the use of glew as all of the OpenGL extensions
are already defined in the GL.h header file. I introduced a macro that
is set to GLEW_VERSION_2_0 for all non-apple systems and defaults to
true on MacOS X. This is safe, since all systems we support have at
least OpenGL 2.0.