improve query editor and dialog
This fixes BZDNG-159 and also improves the general user experience
when using the query editor: Instead of basing the interface directly
on the query flags, we put some abstraction in between and use more
descriptive terms. Also, use a timer to delay the display of error
messages in case something is wrong with the query. This is much
easier on the eyes.

git-svn-id: https://dng.biozentrum.unibas.ch/svn/openstructure/trunk@2663 5a81b35b-ba03-0410-adc8-b2c5c5119f08