make AttachView stricter
The method now checks for agreement of the one letter code in the structure
and the character in the sequence. If there is a mismatch, AttachView throws
an IntegrityError telling you where in the sequence the problem was
detected. The one letter code of the residue and the character in the
sequence agree, if either one of the is X or they are identical.