file type dialog: double click on file type triggers loading
this seemingly simple change required a rather gross hack to avoid a crash when
the file type dialog is destroyed. Currently the FileTypeDialog is allocated on
the heap an leaks. This is not nice, but since the file type dialog is very
rarely used, this seems to be something we can live with for the moment. I've
opened ticket BZDNG-192 to track the issue.