fix segfault for BZDNG-242
The cause for this is actually a bug in Boost.Python that triggers when
arg("arg")... is used in the export and the number of arg("") arguments
doesn't match the number of argument of the function. When one then uses
a wrong kwarg, the program crashes. I'll submit a bug report to the
Boost.Python on this issue. I fear that there are many more cases in
the pymod export where we are using arg("") as we did for the query.