fix AddCompound for residues with atoms not listed in order of increasing ordinals
The issues manifests when importing large "composite" compounds,
e.g. a ligand derived from several amino acids with the atoms not
listed in order of increasing ordinals. This was not a problem up
until now, since the components.cif dictionary was always (as
far as I know) listing the atoms ordered by ordinal.