fixes and modifications to gfx.Gradient; pooled gfx py tests into one file
In Python, gfx.Gradient can now be initialized with a dictionary; for each
key: value pair, the key is a float between 0 and 1, and the value is either
a float triplet or a gfx.Color object.

In C++, Stop and StopList were moved into Gradient to avoid namespace
pollution. In Python, they were renamed to GradientStop and GradientStopList.
As a consequence, gui/pymod/scene/gradient_editor_widget.py has had a rename
of its GradientStop to MyGradientStop.