add util to run python unittests
adding the function RunTests to each python unittest allows that the
test is executed either with the default TestRunner or an XMLTestRunner
for producing an XML output file. With no command line argument, the
default Runner is used (same behaviour as previously), with 'xml' as
the first argument, the XMLTestRunner is used.