let minimal OST compile with barebone CentOS 5.5
A minimalistic OST (no info, no gfx, no gui, no img) compiles
and links with a default CentOS installation. Only eigen2 is
required as an additional dependency. But since that's header
only library that's not a problem for bundling a standalone OST.

Tested on CentOS 5.5 with boost 1.33.1 and MacOS X with boost
1.46.