prefer versioned python binaries to non-versioned ones
On many systems, there are several parallel python
installations. There is usually a "python" symlink that
points to the currently active python version. That might
not be the one we need. Instead of picking up that one, we
now search for python$VERSION and as a last resort use
python.