changed data observer to use ImageHandle instead of data reference changed data viewer to delete on close
The data observer now stores an image handle instead of a data reference.
this means that no external image handle has to be stored to keep a data
viewer open. As a consequence the data viewer window gets deleted, once it
is closed to allow deallocation of the image memory if necessary.
A guarded QPointer<DataViewer> in the proxy prevents access to the
deleted DataViewer from python.

These changes will allow further downstream refactoring of the ImageHandle
and DataViewer (i.e. merge of Data and ImageHandle classes and use of
the full ImageHandle interface in the DataViewer)