reimplementation of PlotBar function
The previous function was hard to understand. Every specified column gets
plotted at one position in the plot. If there are several rows, several
row indices are passed as argument respectively, the values of the
resulting column get plotted in parallel at this position. Note, that
the number off rows is limited to 7.