fix context menu/main menu for newer Qt versions
the ownership semantics of sip/PyQt changed between versions. Object
which were previously staying in memory are now garbage collected.
Event handlers associated with these objects were then removed.
The fix presented in this patch is rather ugly. We store references to
the objects inside the _Init* functions. Ideally, these types of hacks
would not be required, but I could not find another way to approach
the problem.