remove 'ost' directory from sys.path
since the ost startup script lives in the same directory as the
top-level ost modules (python{version}/site-packages/ost), the
directory got automatically added to sys.path. That's bad for
several reasons I don't want to go in right now (see discussion
on the mailing list). Instead of passing the script as a script
name to execute, we now use execfile with the startup script.
This circumvents addition of the directory to the path and
makes Gabriel smile again.