iterative superposition
After an initial superposition, subsets get built using a distance
threshold. These subsets get superposed iteratively until a maximum
number of iterations is reached or the transformation matrix
between two superposition steps converges to the identity matrix.
This procedure usually converges to the largest group of superposable
atoms and is usefull for flexible domains. An example is the
superposition of 1AKE and 4AKE.