fixed bug with possible naming conflicts
When a chain has been inserted in an entity with a ChainHandle as
argument, the name of the given ChainHandle has been used.
The chain then got subsequently renamed to the given name.
This can cause problems, if a chain with the same name as the
input ChanHandle is already present.