Additional flag for topology creator
In the topology creator, the input entity has been renamed to the
naming defined in the forcefield. After setting up the topology,
the naming does not matter anymore, as long as there are no
manipulations on the atoms. With the additional flag, the topology
creator can rename the entity back to pdb standard, as soon as
the whole parametrization magic took place.