properly remove torsion handles from residues
When an atom gets deleted, all related torsions get deleted from
the entities TorsionImplMap. The ResidueHandles keep their own vector
of the same TorsionHandles. If an atom gets deleted, only the internal
torsion list of the corresponding residue got updated. This had to be
extended to all residues containing torsions related to the deleted
atom.