Solvent Accessibility after Lee and Richards
It's basically a clone of NACCESS. The current function interface
mimics the interface from the naccess binding.

The problem with NACCESS is the size limitation of the input. It also
allocates tons of memory it doesn't need. The C++ implementation should
be much more careful with the memory and thus allows has much higher
limits for the input size. It's also faster due to less io (no dump
of pdb file on HD as input for NACCESS). On my machine, the speed
increase is roughly 2x.

The results have been checked for equality with NACCESS on a large
set of structures.

Fails have been observed in:

- structures containing hydrogens:
Bug in NACCESS... Example: the hydrogen hanging at the CG2 of THR
are called HG21, HG22 and HG23 => four letters and the atom name
already starts at position 13 in the according line in the PDB file.
NACCESS checks if position 14 is either H or D to detect a hydrogen.
This fails in this case (letter at pos 14 is G). Since its an "unknown"
atom, the hydrogen gets treated as carbon with VdW radius 1.80.
Example in THR.383.A in pdb entry 4DKK

- structures containing alt locations
Maybe different handling of which alternative is actually active.
Has to be further evaluated.