avoid widget showing up twice
QT5 shows a widget twice when you set a parent at initialization
and reset the parent later. This is exactly what happens when a
widget is initialized and passed to AddWidget in the gosty app.
The solution is to initialize them without a parent and let
the AddWidget function taking care of it.