SCHWED-4266 Solvent Accessibility failure for deep view models
Atoms that do not actually exist are still added in deep view. The problem is
that they're added at position (9999.99, 9999.99, 9999.99). Internally we use
a grid of equally sized cubes to efficiently lookup up positions closeby.
Extreme locations lead to an extremely large grid with most of the cubes being
unoccupied. The simple solution here is to just increase the size of the
cubes. The larger the cubes, the slower the code. However, in real world
scenarios their size won't be inreased.